desafio/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
filelock==3.10.7
platformdirs==3.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
virtualenv==20.21.0